increase privacy and set authentication step for accessing

+++ conterllers/contactController.js
@@ -6,7 +6,7 @@ const Contact = require('../models/contactModel');
 //@acesses private
 
 const getContacts = asyncHnadler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
@@ -15,14 +15,24 @@ const getContacts = asyncHnadler(async (req, res) => {
 //@acesses private
 
 const getContact = asyncHnadler(async (req, res) => {
+
     const contact = await Contact.findById(req.params.id);
+    
     if (contact){
-        res.status(200).json(contact);
+        if (contact.user_id.toString() == req.user.id){
+            res.status(200).json(contact);
+        }
+        else{
+            res.status(400);
+            throw new Error("You are not validate!")
+        }      
     }
     else{
         res.status(404);
         throw new Error("Contact not found!")
     }   
+     
+  
 });
 
 //@desc make new contact
@@ -31,9 +41,9 @@ const getContact = asyncHnadler(async (req, res) => {
 
 const createContact = asyncHnadler(async (req, res) => {
     const {name, email, phone} = req.body;
-    console.log(req.body);
     if (name && email && phone){
         const contact = await Contact.create({
+            user_id: req.user.id,
             name,
             email,
             phone
@@ -53,18 +63,24 @@ const createContact = asyncHnadler(async (req, res) => {
 const editContact = asyncHnadler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (contact){
-        const updatedContact = await Contact.findByIdAndUpdate(
-            req.params.id,
-            req.body,
-            { new : true }
-        );
-        res.status(200).json(updatedContact);
+        if (contact.user_id.toString() == req.user.id){
+            const updatedContact = await Contact.findByIdAndUpdate(
+                req.params.id,
+                req.body,
+                { new : true }
+            );
+            res.status(200).json(updatedContact);
+        }
+        else{
+            res.status(400);
+            throw new Error("You are not validate to do that");
+        }
     }
     else{
         res.status(404);
         throw new Error("Contact not found!")
     }
-
+    
 });
 
 //@desc delete specefic contact
@@ -74,10 +90,15 @@ const editContact = asyncHnadler(async (req, res) => {
 const deleteContact = asyncHnadler(async (req, res) => {
    
     const contact = await Contact.findById(req.params.id);
-    
     if (contact){
-        await Contact.deleteOne({ _id: req.params.id });
-        res.status(200).json(contact);
+        if (contact.user_id.toString() == req.user.id){
+            await Contact.deleteOne({ _id: req.params.id });
+            res.status(200).json(contact);
+        }
+        else{
+            res.status(400);
+            throw new Error("You are not validate to do that");
+        } 
     }
     else{
         res.status(404);
